% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_func.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Title}
\usage{
cosine_distance(a, b)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector of same length.}
}
\value{
The cosine_distance of a,b.
}
\description{
Title
}
