% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_module.R
\name{filter_module}
\alias{filter_module}
\title{Title}
\usage{
filter_module(
  object,
  module,
  sample_name,
  read_dir,
  out_dir = NULL,
  cluster_col,
  sd_thres = 0.01,
  mean_thres = NA,
  pct_thres = NA,
  save_fig = FALSE,
  number_per_fig = 9,
  fig_dir = NULL,
  save_fig_spatial = FALSE,
  spot_size,
  merge_module = TRUE
)
}
\arguments{
\item{object}{A seurat object}

\item{module}{A list,the result of get_module}

\item{sample_name}{A string indicating the sample name ,corresponding to the input of generate_module&score_module.}

\item{read_dir}{The output directory path of fuction score_module.}

\item{out_dir}{The output directory of filtered module.}

\item{cluster_col}{A string corresponding to one colname of Seurat object metadata,indicating the cell cluster assignment.}

\item{sd_thres}{A non-negative double,indicating the threshold of the difference between two standard deviations.}

\item{mean_thres}{A non-negative double,indicating the threshold of the difference between two mean values.}

\item{pct_thres}{A non-negative double,indicating the threshold of the proportion of cells with program score greater than zero.}

\item{save_fig}{A boolean variable that specifies whether to save the visualization of the program score.}

\item{number_per_fig}{A positive integer that specifies the number of programs saving in one figure.}

\item{fig_dir}{A string indicating the directory path of the figures.}

\item{save_fig_spatial}{A boolean variable that specifies whether to save the spatial visualization of the program score.FALSE by default,only saving the UMAP visualization of the program score,TRUE,saving the spatial dimplot of the program score.}

\item{spot_size}{A positive double,indicating the spot size of the spatial dimplot.}

\item{merge_module}{A boolean variable that specifies whether to merge modules generated from the same clusters.}
}
\value{
A list,'chosen' is a dataframe,storing the information of the programs kept and their corresponding clusters;'obj' is a Seurat object, in which the metadata has been updated with the scoring values from the SSpMosaic programs;'sd'/'fc'/'pct' are lists named after clusters,where each item is a vector named after gene programs where each item is the program's metric value.
}
\description{
Title
}
