% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_propagation.R
\name{get_individual_heats_matrix}
\alias{get_individual_heats_matrix}
\title{Title}
\usage{
get_individual_heats_matrix(
  nam_or_graph,
  alpha = 0.5,
  conserve_heat = TRUE,
  weighted = FALSE
)
}
\arguments{
\item{nam_or_graph}{A igraph object or a normalized_adjacency_matrix}

\item{alpha}{A double number,between 0  and 1}

\item{conserve_heat}{A boolean variable that specifies whether to maintain heat conservation during network propagation.}

\item{weighted}{A boolean variable that specifies whether the biological network is weighted.}
}
\value{
A matrix,individual_heats_matrix
}
\description{
Title
}
