% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_propagation.R
\name{get_normalized_adjacency_matrix}
\alias{get_normalized_adjacency_matrix}
\title{Title}
\usage{
get_normalized_adjacency_matrix(graph, conserve_heat = TRUE, weighted = FALSE)
}
\arguments{
\item{graph}{A igraph object,biological network,such as PPI network.}

\item{conserve_heat}{A boolean variable that specifies whether to maintain heat conservation during network propagation.}

\item{weighted}{A boolean variable that specifies whether the biological network is weighted.}
}
\value{
A matrix,normalized adjacency matrix
}
\description{
Title
}
