% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_propagation.R
\name{network_propagation}
\alias{network_propagation}
\title{Title}
\usage{
network_propagation(individual_heats_matrix, nodes, seed_genes)
}
\arguments{
\item{individual_heats_matrix}{A matrix,the result of function get_individual_heats_matrix}

\item{nodes}{A character vector represents all the genes included in the biological network, and the order should be consistent with the individual_heats_matrix.}

\item{seed_genes}{A character vector represents the genes included in the gene program}
}
\value{
A numeric vector,the result of network_propagation
}
\description{
Title
}
