% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_module.R
\name{new_ModuleConnectivity}
\alias{new_ModuleConnectivity}
\title{Title}
\usage{
new_ModuleConnectivity(
  seurat_obj,
  group.by = NULL,
  group_name = NULL,
  corFnc = "bicor",
  corOptions = "use='p'",
  harmonized = TRUE,
  assay = NULL,
  slot = "data",
  sparse = TRUE,
  reassign_modules = TRUE,
  wgcna_name = NULL,
  ...
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{group.by}{column in seurat_obj@meta.data containing grouping info, ie clusters or celltypes}

\item{group_name}{name of the group(s) in group.by to use for kME calculation}

\item{corFnc}{character string specifying the function to be used to calculate co-expression similarity.}

\item{corOptions}{character string specifying additional arguments to be passed to the function given by corFnc}

\item{harmonized}{logical determining whether to use harmonized MEs for kME calculation}

\item{assay}{Assay in seurat_obj containing expression information}

\item{slot}{Slot in seurat_obj, default to normalized 'data' slot}

\item{sparse}{logical indicating whether or not to run the correlation using a sparse matrix}

\item{reassign_modules}{logical indicating whether or not to reassign genes to different co-expression modules if their kME value in the assigned module is negative}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
A Seurat object with the kMEs computed for the selected wgcna experiment
}
\description{
Title
}
