% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_module.R
\name{new_ReassignModules}
\alias{new_ReassignModules}
\title{Title}
\usage{
new_ReassignModules(
  seurat_obj,
  harmonized = TRUE,
  features = NULL,
  new_modules = NULL,
  ignore = FALSE,
  wgcna_name = NULL
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{harmonized}{logical indicating whether or not to use harmonized MEs}

\item{features}{character vector containing features for manual module reassignment}

\item{new_modules}{character vector containing modules to reassign the genes}

\item{ignore}{logical indicating whether or not to ignore error message about reassigning non-grey features}

\item{wgcna_name}{The name of the hdWGCNA experiment in the seurat_obj@misc slot}
}
\value{
A Seurat object with the an updated modules table for the selected wgcna experiment
}
\description{
Title
}
