% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_module.R
\name{score_module}
\alias{score_module}
\title{Title}
\usage{
score_module(
  module,
  sample_name,
  read_dir,
  out_dir = NULL,
  cluster_col,
  nbin = 2
)
}
\arguments{
\item{module}{A list,the result of get_module,the input sample_name should be consistent.}

\item{sample_name}{A string indicating the sample name ,corresponding to the input of generate_module.}

\item{read_dir}{The output directory path of fuction generate_module.}

\item{out_dir}{The output directory of module score result.}

\item{cluster_col}{A string corresponding to one colname of Seurat object metadata,indicating the cell cluster assignment.}

\item{nbin}{Number of bins of aggregate expression levels for all analyzed features used in Seurat::AddModuleScore}
}
\value{
This fuction has no return.
}
\description{
Title
}
