% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.R
\name{SSpMosaic_decon}
\alias{SSpMosaic_decon}
\title{Title}
\usage{
SSpMosaic_decon(
  seurat_object,
  meta_moudule_genes,
  normalize_method = "zscore",
  assay = "Spatial",
  layer = "counts",
  slot = "counts"
)
}
\arguments{
\item{seurat_object}{A seurat object of spatial transcriptomics data.}

\item{meta_moudule_genes}{A list named after SSpMosaic meta modules,where each item is a character array storing the gene set corresponding to the meta module.}

\item{normalize_method}{A string representing the normalization method, which should be selected from ‘none’, ‘zscore’, or ‘min-max’.}

\item{assay}{Name of assay to pull the gene expressiong data from.}

\item{layer}{Name of layer to fetch data from,for Seurat v5 only.}

\item{slot}{Name of slot to fetch data from,for Seurat v4 only.}
}
\value{
A Seurat object, in which the metadata has been updated with the scoring values from the SSpMosaic meta module, assigning labels to each spot.
}
\description{
Title
}
