% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.R
\name{SSpMosaic_meta_integration.Seurat}
\alias{SSpMosaic_meta_integration.Seurat}
\title{Title}
\usage{
SSpMosaic_meta_integration.Seurat(
  object,
  col_batch,
  col_SSpMosaic_cluster = "SSpMosaic_cluster",
  reduction_use = "pca",
  reduction_dimention = 1:30,
  seed = 42,
  neighbors_within_modality = 3,
  neighbors_across_modality = 3,
  graph_name = "SSpMosaic",
  n_meta_neighbors,
  batch_group_by_modality = NULL,
  calculate_umap = T
)
}
\arguments{
\item{object}{A Seurat object}

\item{col_batch}{one of the colnames of the object's metadata,storing the batch information of the cells}

\item{col_SSpMosaic_cluster}{one of the colnames of the object's metadata,storing the batch information of the cells}

\item{reduction_use}{a string indicating the name of the reductio result used for calculating distance}

\item{reduction_dimention}{A vector of positive integer indicating the PCs used for calculating distance}

\item{seed}{a integer,random seed}

\item{neighbors_within_modality}{A positive integer,the number of cell neighbors from the same modality/species}

\item{neighbors_across_modality}{A positive integer,the number of cell neighbors from the different modality/species}

\item{graph_name}{a string indicating the name used to save the SSpMosaic integration result}

\item{n_meta_neighbors}{A positive integer indicating the number of meta neighbors.}

\item{batch_group_by_modality}{A list named by modality/species names, containing the batch names belonging to this modality/species.}

\item{calculate_umap}{A boolean variable that specifies whether to compute the UMAP dimensionality reduction results after integration.}
}
\value{
A Seurat object after integration
}
\description{
Title
}
