% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{SSpMosaic_sc_annotation}
\alias{SSpMosaic_sc_annotation}
\title{Title}
\usage{
SSpMosaic_sc_annotation(
  object,
  meta_module_genes,
  cluster_col,
  sd_thres = 0,
  mean_thres = 0,
  annotation_name = "annotation"
)
}
\arguments{
\item{object}{A Seurat object containing unsupervised clustering results}

\item{meta_module_genes}{A list named after SSpMosaic meta modules,where each item is a character array storing the gene set corresponding to the meta module.}

\item{cluster_col}{A string corresponding to one colname of Seurat object metadata,indicating the cell cluster assignment.}

\item{sd_thres}{A non-negative double,indicating the threshold of the difference between two standard deviations.}

\item{mean_thres}{A non-negative double,indicating the threshold of the difference between two mean values.}

\item{annotation_name}{A string,indicating the column in meta.data to store the annotation information.}
}
\value{
A list,'object' is a Seurat object, in which the metadata has been updated with the SSpMosaic annotation;'mean_dif' is a list,in which the differences between mean values are stored;'sd_dif' is a list,in which the differences between standard deviations values are stored
}
\description{
Title
}
