% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{weighted_score}
\alias{weighted_score}
\title{Title}
\usage{
weighted_score(
  object,
  meta_module_genes,
  assay = "RNA",
  layer = "counts",
  slot = "counts",
  normalize_method = "none",
  anno = F
)
}
\arguments{
\item{object}{A seurat object of single cell data}

\item{meta_module_genes}{A list named after SSpMosaic meta modules,where each item is a character array storing the gene set corresponding to the meta module.}

\item{assay}{Name of assay to pull the gene expressiong data from.}

\item{layer}{Name of layer to fetch data from,for Seurat v5 only.}

\item{slot}{Name of slot to fetch data from,for Seurat v4 only.}

\item{normalize_method}{A string representing the normalization method, which should be selected from ‘none’, ‘zscore’, or ‘min-max’.}

\item{anno}{A boolean variable that specifies whether to annotate the cells with the highest-scoring meta-module.}
}
\value{
A Seurat object, in which the metadata has been updated with the scoring values from the SSpMosaic meta module, assigning labels to each spot.
}
\description{
Title
}
